requirements.txt
asgiref==3.8.1
async-timeout==4.0.2
blinker==1.6.2
branca==0.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudinary==1.40.0
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.0.0
django-environ==0.11.2
django-mathfilters==1.0.0
django-phonenumber-field==7.3.0
django-redis==5.3.0
djangorestframework==3.14.0
Flask==2.3.3
folium==0.14.0
geographiclib==2.0
geopy==2.4.0
greenlet==2.0.2
gunicorn==22.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
packaging==24.0
pandas==2.1.0
parse-rest==0.10.2013
phonenumbers==8.13.31
Pillow==9.5.0
psycopg2-binary==2.9.9
pysqlite3==0.5.0
python-dateutil==2.8.2
pytz==2023.3
query==0.1.4
redis==4.5.5
requests==2.31.0
six==1.10.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
stripe==7.0.0
typing_extensions==4.6.0
tzdata==2023.3
urllib3==2.0.4
uuid==1.30
uvicorn==0.30.1
Werkzeug==2.3.7
whitenoise==6.6.0
